refactor: simplify deps API
