+++ package.json
@@ -21,9 +21,6 @@
       ],
       "core": [
         "./build/core.d.ts"
-      ],
-      "internals": [
-        "./build/internals.d.ts"
       ]
     }
   },
@@ -33,7 +30,6 @@
     "./experimental": "./build/experimental.js",
     "./cli": "./build/cli.js",
     "./core": "./build/core.js",
-    "./internals": "./build/internals.js",
     "./package.json": "./package.json"
   },
   "bin": {
